cr
Former-commit-id: 3fae21a3d793679dcab7912d27874c1c3de7842b

+++ WalletManager/ELWalletManager.h
@@ -167,7 +167,7 @@ typedef std::map<String, IDidManager*> DIDManagerMap;
 //参加投票
 -(NSInteger)RegisterProducerWithMainchainSubWallet:(IMainchainSubWallet*)ELA With:(FLJoinVoteInfoModel*)model;
 -(NSInteger)UpdateProducerWithMainchainSubWallet:(IMainchainSubWallet*)ELA With:(FLJoinVoteInfoModel*)model;
--(BOOL)useCRMainchainSubWallet:(NSString*)CRmainchainSubWalletId ToVote:(NSDictionary*)publicKeys tickets:(NSInteger)stake pwd:(NSString*)pwd isChangeVote:(BOOL)change;
+-(BOOL)useCRMainchainSubWallet:(NSString*)CRmainchainSubWalletId ToVote:(NSDictionary*)publicKeys tickets:(double)stake pwd:(NSString*)pwd isChangeVote:(BOOL)change;
 -(BOOL)useMainchainSubWallet:(NSString*)mainchainSubWalletId ToVote:(NSArray*)publicKeys tickets:(double)stake pwd:(NSString*)pwd isChangeVote:(BOOL)change;
 -(PluginResult *)SignReadOnlyToVote:(invokedUrlCommand *)command;
 -(PluginResult *)HowSignToVote:(invokedUrlCommand *)command;
